Move LocksAndMetadata
