+++ atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/keyvalue/api/watch/LocksAndMetadata.java
@@ -1,0 +1,37 @@
+/*
+ * (c) Copyright 2023 Palantir Technologies Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.palantir.atlasdb.keyvalue.api.watch;
+
+import com.palantir.lock.LockDescriptor;
+import com.palantir.lock.watch.LockRequestMetadata;
+import com.palantir.logsafe.Unsafe;
+import java.util.Optional;
+import java.util.Set;
+
+@Unsafe
+@org.immutables.value.Value.Immutable
+public interface LocksAndMetadata {
+    @org.immutables.value.Value.Parameter
+    Set<LockDescriptor> lockDescriptors();
+
+    @org.immutables.value.Value.Parameter
+    Optional<LockRequestMetadata> metadata();
+
+    static LocksAndMetadata of(Set<LockDescriptor> lockDescriptors, Optional<LockRequestMetadata> metadata) {
+        return ImmutableLocksAndMetadata.of(lockDescriptors, metadata);
+    }
+}
